backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==2.3.3
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
plyer==2.1.0
psutil==7.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pywin32==311; platform_system == "Windows"
Werkzeug==2.3.8
gunicorn==22.0.0
flask-jwt-extended==4.4.4
